Fix issue with ExtractSubMPX
Lots of dubious assumptions meant that ExtractSubMPX was borken.
Have removed this function and replaced uses with ExtractMPOBlock
which does everything that was needed. This fixes a segfault in
TEBD/iTEBD that had been introduced by the dynamical correlations
fix.

+++ arpack_interface.hpp
@@ -20,7 +20,7 @@ namespace arpack {
   typedef long long arpack_int; 
 #else
   /** Else use normal int type */
-  typedef int arpack_int;
+  typedef int64_t arpack_int;
 #endif
 
   /** external arpack function */
